requirements.txt
# Các thư viện xử lý tài liệu
python-docx==0.8.11
pdfminer.six==20221105
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.5.0

# Các thư viện cho vector embedding và retrieval
# Đã loại bỏ sentence-transformers để tránh xung đột với huggingface-hub
transformers>=4.31.0,<4.38.0
huggingface-hub>=0.16.4
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.0.1

# Các thư viện cho LLM và API
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8

# Các thư viện hỗ trợ
tqdm==4.65.0
tokenizers>=0.13.3 